Unix: default file extension of new files: `.md`
For Unix-like systems this defaults to `.md` because all the
listed file editors (see `APP_ARGS_EDITOR`) support it. The
Windows default is `.txt` to ensure that the _Notepad_ editor can
handle these files properly.

The Windows ecosystem does not support `.md` well enough yet. In a
default Windows installation the only file editor available
is Notepad. Unfortunately Notepad does not register the `.md`
extension. This means, when you send a `.md` note file to a
colleague, who does not have Tp-Note installed, can not open
the file.