Viewer: watch the whole note's directory
In theory watching only `file` is enough. Unfortunately some file
editors do not modify files directly. They first rename the existing
file on disk and then  create a new file with the same filename. As
a workaround, we watch the whole directory where the file resides.
False positives, there could be other changes in this directory
which are not related to `file`, are detected, as we only trigger
the rendition to HTML when `debounced_event.path` corresponds to our
watched file.