Fix edge case: clipboard has only sort-tag chars
This fixes a minor regression introduced with Tp-Note v1.21.10: Note files,
created with title fields containing only sort tag characters, need an
`extra_separator` between in the sort-tag and the title in their filename. This
caused cyclic filename change when synchronizing the note's metadata and the
filename. Now, an extra-separator is inserted in this case also.