Do not shorten shorthand autolinks, e.g.:`<tpnote:3.0>`
A note file can be referenced either with a complete autolink e.g.
`<tpnote:3.0-my_note.md>` containing the full filename, or with
it's shorthand, e.g. `<tpnote:3.0>`, containing only the sort-tag part.
When a complete autolink is rendered to HTML, only the file stem. e.g.
`3.0-ny_note` is displayed. The above shorthand form is displayed
as `3.0`.

This commit fixes a corner case in the renderer, when a shorthand
autolink contains a dot `.` link in `<tpnote:3.0>`.