Viewer: roll back file watcher workaround
Some file editors do not change the file on save, but rename the current
version and place it with a new file at the same path. This case was not
properly detected by earlier versions of Notify. The workaround was to
watch the whole directory instead. Recent tests show, that this
shortcoming seems to be overcome. This commit rolls back the workaround.